backend/requirements.txt
# ── Core ──────────────────────────────────────
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
python-dotenv==1.0.1

# ── Data ──────────────────────────────────────
pandas==2.2.3
pyarrow==19.0.1
numpy==1.26.4

# ── Embeddings & FAISS ────────────────────────
sentence-transformers==2.7.0
faiss-cpu==1.9.0.post1
torch>=2.0.0

# ── BM25 Sparse Retrieval ────────────────────
rank-bm25==0.2.2

# ── LangChain + LangGraph ────────────────────
langchain==0.3.25
langchain-core==0.3.59
langchain-community==0.3.23
langchain-groq==0.3.2
langgraph==0.3.34

# ── Misc ──────────────────────────────────────
tqdm>=4.66.0